backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
httpx==0.27.0
pydantic==2.7.0
python-dotenv==1.0.1
tree-sitter==0.23.2
tree-sitter-python==0.23.6
tree-sitter-typescript==0.23.2
tree-sitter-javascript==0.23.1
networkx==3.4.2
gitpython==3.1.43
faiss-cpu==1.9.0
sentence-transformers==3.3.1
watchdog==6.0.0
aiofiles==24.1.0
numpy==1.26.4
langgraph==0.2.55
langchain-core==0.3.28
sse-starlette==2.1.3
